I6 Move User graphql logic to User model
