+++ node_server/src/main/graphQL/schema.ts
@@ -1,8 +1,7 @@
 import { GraphQLSchema } from 'graphql';
-import { composeWithMongoose } from 'graphql-compose-mongoose';
 import mongoose from 'mongoose';
 import { schemaComposer } from 'graphql-compose';
-import { createUserModel } from '../models/user';
+import { createUserTC, addUserFieldsToSchema } from '../models/user';
 
 /**
  * Creates the GraphQL schema for the project from the mongoose models.
@@ -13,16 +12,9 @@ import { createUserModel } from '../models/user';
 export default function createGraphQLSchema(
   db: typeof mongoose
 ): GraphQLSchema {
-  const User = createUserModel(db);
-  const UserTC = composeWithMongoose(User);
+  const UserTC = createUserTC(db);
 
-  schemaComposer.Query.addFields({
-    userMany: UserTC.getResolver('findMany'),
-  });
-
-  schemaComposer.Mutation.addFields({
-    userCreateOne: UserTC.getResolver('createOne'),
-  });
+  addUserFieldsToSchema(UserTC, schemaComposer);
 
   return schemaComposer.buildSchema();
 }
